Conditionals for missing paths + fix find min/max depth position

+++ lib/kitchen/provisioner/formula-fetch.sh
@@ -64,19 +64,23 @@ function linkFormulas() {
   SALT_ROOT=$1
   SALT_ENV=${2:-/usr/share/salt-formulas/env}
 
-  # form git, development versions
-  GIT_FORMULA_LIST=$(find "$SALT_ENV"/_formulas -maxdepth 1 -mindepth 1 -type d)
-  for formula in $GIT_FORMULA_LIST
-  do
-    name=$(basename "$formula")
-    ln -fs "$formula/$name" "$SALT_ROOT/$name"
-    find "$formula" -maxdepth 1 -mindepth 1 -type d |grep -E "_(modules|states|grains|renderers|returners)" | xargs -I{} \
-      basename {}| xargs -I{} cp -rs "$formula"/{} "$SALT_ROOT"/
-  done
+  if [[ -e "$SALT_ENV" ]]; then
+    # form git, development versions
+    if [[ -e "$SALT_ENV/_formulas" ]]; then
+      GIT_FORMULA_LIST=$(find "$SALT_ENV"/_formulas -maxdepth 1 -mindepth 1 -type d)
+      for formula in $GIT_FORMULA_LIST
+      do
+        name=$(basename "$formula")
+        ln -fs "$formula/$name" "$SALT_ROOT/$name"
+        find "$formula" -maxdepth 1 -mindepth 1 -type d |grep -E "_(modules|states|grains|renderers|returners)" | xargs -I{} \
+          basename {}| xargs -I{} cp -rs "$formula"/{} "$SALT_ROOT"/
+      done
+    fi
 
-  # form pkgs
-  find "$SALT_ENV" -path "*_formulas*" -prune -o -name "*" -maxdepth 1 -mindepth 1 -type d -print0| xargs -0 -n1 basename | xargs -I{} \
-    ln -fs "$SALT_ENV"/{} "$SALT_ROOT"/{};
+    # form pkgs
+    find "$SALT_ENV" -maxdepth 1 -mindepth 1 -path "*_formulas*" -prune -o -name "*" -type d -print0| xargs -0 -n1 basename | xargs -I{} \
+      ln -fs "$SALT_ENV"/{} "$SALT_ROOT"/{};
+  fi
 
 }
